% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHI.R
\name{CHI}
\alias{CHI}
\title{Convex Hierarchical Portfolio}
\usage{
CHI(
  sigma,
  mu = NULL,
  meta_loss = c("MaxDiv", "ERC"),
  UB = NULL,
  LB = NULL,
  groups = NULL,
  group.UB = NULL,
  group.LB = NULL,
  gamma = 0,
  max_tilt = 1,
  groups_mat = NULL,
  verbose = F,
  ...
)
}
\arguments{
\item{sigma}{a \eqn{(N \times N)}{(N x N)} covariance matrix.}

\item{mu}{a \eqn{(N \times 1)}{(N x 1)} vector of estimated returns.}

\item{meta_loss}{a loss function of the most diversified hierarchical allocation graph.}

\item{UB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of upper bound weight constraints.}

\item{LB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of lower bound weight constraints.}

\item{groups}{vector of group IDs. The names of the vector must be identical to the asset names.}

\item{group.UB}{scalar or \eqn{(N_groups\times 1)}{(N_groups x 1)} vector of upper bound group constraints.}

\item{group.LB}{scalar or \eqn{(N_groups\times 1)}{(N_groups x 1)} vector of lower bound group constraints.}

\item{gamma}{risk aversion parameter. Default: \code{gamma = 0} returns the minimum variance portfolio.}

\item{max_tilt}{maximum percentage reduction in the effective number of assets. Default: \code{max_tilt = 1} (no restriction).}

\item{groups_mat}{Group constraints passed to MV.}

\item{verbose}{Set to FALSE by default. If True, it returns the weights vector and the covariance matrix.}

\item{...}{arguments passed to \code{cluster::agnes} method.}
}
\value{
A \eqn{(N \times 1)}{(N x 1)} vector of optimal portfolio weights.
}
\description{
Computes the optimal CHI-MVO portfolio with full investment, weight and group constraints.
}
\details{
The argument \code{sigma} is a covariance matrix.

Hierarchical clustering is performed using the \code{cluster}-package. If
\code{cluster_method == 'DIANA'}, the function \code{cluster::diana} is used
to compute a cluster dendrogram, otherwise the function \code{cluster::agnes(., method = cluster_method)}
is used. Default is single-linkage agglomerative nesting.

The argument \code{meta_loss} represents the loss function used to optimize the most diversified hierarchical allocation graph.
The optimized hierarchy is used to filter \code{sigma} and \code{mu}. If the filtered covariance matrix is used in a
mean variance portfolio optimizer, a CHI portfolio is constructed.
}
\examples{
# Load returns of assets or portfolios
data("Industry_10")
rets <- Industry_10
sigma <- cov(rets)
CHI(sigma, UB = 0.15)

}
\author{
Johann Pfitzinger
}
