% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiPlot.R
\name{chiPlot}
\alias{chiPlot}
\title{Plot the dendrogram of a CHI portfolio}
\usage{
chiPlot(
  sigma,
  mu = NULL,
  meta_loss = c("MaxDiv", "ERC"),
  UB = NULL,
  LB = NULL,
  groups = NULL,
  group.UB = NULL,
  group.LB = NULL,
  gamma = 0,
  max_tilt = 1,
  max_leaf_size = 3,
  ymax = NULL,
  horiz = FALSE,
  ...
)
}
\arguments{
\item{sigma}{a \eqn{(N \times N)}{(N x N)} covariance matrix.}

\item{mu}{a \eqn{(N \times 1)}{(N x 1)} vector of estimated returns.}

\item{meta_loss}{a loss function of the most diversified hierarchical allocation graph.}

\item{UB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of upper bound weight constraints.}

\item{LB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of lower bound weight constraints.}

\item{groups}{vector of group IDs. The names of the vector must be identical to the asset names.}

\item{group.UB}{scalar or \eqn{(N_groups\times 1)}{(N_groups x 1)} vector of upper bound group constraints.}

\item{group.LB}{scalar or \eqn{(N_groups\times 1)}{(N_groups x 1)} vector of lower bound group constraints.}

\item{gamma}{risk aversion parameter. Default: \code{gamma = 0} returns the minimum variance portfolio.}

\item{max_tilt}{maximum percentage reduction in the effective number of assets. Default: \code{max_tilt = 1} (no restriction).}

\item{max_leaf_size}{maximum size of the leaf nodes. Default is \code{max_leaf_size=3}.}

\item{ymax}{upper limit for y axis}

\item{horiz}{rotate the dendrogram}

\item{...}{arguments passed to \code{cluster::agnes} method.}
}
\value{
A plotted dendrogram.
}
\description{
Plots the dendrogram of a CHI portfolio.
}
\details{
The dendrogram is generated using hierarchical clustering and modified
so that the height differential between any two splits is the shrinkage weight of
the lower split (ranging between \code{0} and \code{1}). With no shrinkage, all shrinkage weights
are equal to \code{1} and the dendrogram has a height of \eqn{p}{p}. With shrinkage
the dendrogram has a height of \eqn{(\kappa \times p)}{(\code{kappa} x p)}.

The leaf nodes are colored to indicate the coefficient sign, with the size indicating
the absolute magnitude of the coefficients.

A color bar on the right indicates the relative contribution of each level to the
coefficient of determination, with darker hues representing a larger contribution.
}
\examples{
# Load returns of assets or portfolios
data("Industry_10")
rets <- Industry_10
sigma <- cov(rets)
chiPlot(sigma, UB = 0.15)

}
\seealso{
\code{CHI} and \code{chiSigma} methods
}
\author{
Johann Pfitzinger
}
