% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Opt.R
\name{General_Opt}
\alias{General_Opt}
\title{General Optimizer Routine (nesting many). See example notation by running data("spec")}
\usage{
General_Opt(
  Optim = c("All", "MV", "MaxSharpe", "MinVol", "MaxDiv", "ERC"),
  sigma,
  mu = NULL,
  UB = NULL,
  LB = NULL,
  groups = NULL,
  group.UB = NULL,
  group.LB = NULL,
  groups_mat = NULL,
  gamma = 0,
  DivRatio = DR,
  rf = 0,
  maxeval = 5000,
  xtol_rel = 1e-10,
  ftol_abs = 1e-10,
  tol = 1e-10
)
}
\arguments{
\item{Optim}{Optimization techniques used: posibilities: c("All", "MV", "MaxSharpe", "MinVol", "MaxDiv", "ERC"). All binds rows if feasible.}

\item{sigma}{a \eqn{(N \times N)}{(N x N)} covariance matrix.}

\item{mu}{a \eqn{(N \times 1)}{(N x 1)} vector of estimated returns.}

\item{UB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of upper bound weight constraint.}

\item{LB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of lower bound weight constraint.}

\item{groups}{vector of group IDs. The names of the vector must be identical to the asset names.}

\item{group.UB}{scalar or \eqn{(N_groups\times 1)}{(N_groups x 1)} vector of upper bound group constraints.}

\item{group.LB}{scalar or \eqn{(N_groups\times 1)}{(N_groups x 1)} vector of lower bound group constraints.}

\item{gamma}{risk aversion parameter. Higher means more risk averse - this is multiplied by cov. Default: \code{gamma = 0}.}

\item{rf}{optional risk free rate (annualised).}

\item{maxeval}{used in nloptr.}

\item{xtol_rel}{used in nloptr.}

\item{ftol_abs}{used in nloptr.}

\item{tol}{used in nloptr.}
}
\value{
A \eqn{(N \times 1)}{(N x 1)} vector of optimal portfolio weights.
}
\description{
Computes a chosen portfolio optimizer with full investment and weight constraints.
}
\details{
The argument \code{sigma} is a covariance matrix.
The MV solution is calculated using \code{quadprog}, others use NLOPTR
}
\examples{
data("spec")
spec$n -> n; spec$gamma -> gamma; spec$sigma -> sigma; spec$mu -> mu; spec$group.LB->group.LB; spec$group.UB -> group.UB;
spec$groups_mat -> groups_mat; spec$groups -> groups; spec$LB -> LB; spec$UB -> UB; spec$assets -> assets; rf = 0;
maxeval = 5000; xtol_rel = 1e-10; ftol_abs = 1e-10; tol = 1e-10
all_opts <- General_Opt( Optim = "All", sigma, mu = mu, UB = UB, LB = LB, groups = groups, group.UB = group.UB, group.LB = group.LB, groups_mat = groups_mat, gamma = gamma)
MV_Roll_gamma <- seq(0.1, 150, length.out = 100) \%>\% as.list() \%>\% map_df(~General_Opt( Optim = "MV", sigma, mu = mu, UB = UB, LB = LB, groups = groups, group.UB = group.UB, group.LB = group.LB, groups_mat = groups_mat, gamma = .))
}
\author{
Nico Katzke
}
