% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRP.R
\name{HRP}
\alias{HRP}
\title{Hierarchical Risk Parity portfolio}
\usage{
HRP(sigma, UB = NULL, LB = NULL, tau = 0, ...)
}
\arguments{
\item{sigma}{a \eqn{(N \times N)}{(N x N)} covariance matrix.}

\item{UB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of upper bound weight constraint.}

\item{LB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of lower bound weight constraint.}

\item{tau}{trade-off between naive (0) or cluster-based (1) tree-splitting (see Details).}

\item{...}{arguments passed to \code{cluster::agnes} method.}
}
\value{
A \eqn{(N \times 1)}{(N x 1)} vector of optimal portfolio weights.
}
\description{
Computes optimal HRP portfolio with full investment and weight constraints.
}
\details{
The argument \code{sigma} is a covariance matrix and does not need to be positive definite.

Hierarchical clustering is performed using the \code{cluster}-package. If
\code{cluster_method == 'DIANA'}, the function \code{cluster::diana} is used
to compute a cluster dendrogram, otherwise the function \code{cluster::agnes(., method = cluster_method)}
is used. Default is single-linkage agglomerative nesting.

By default, HRP constructs a balanced graph by bisecting each cluster at the
central location \eqn{(tau = 0)}{(tau = 0)}. Alternatively it is possible to split along
the actual cluster dendrogram \eqn{(tau = 1)}{(tau = 1)}, or to trade-off the objectives
by setting \eqn{0 < tau < 1}{0 < tau < 1}. In the latter case, \code{tau} is the proportion
of nodes on the left and right of the central split from which to select the
optimal splitting location (i.e. split at the optimal location that is at
most \eqn{tau\times x N/2}{tau x N/2} elements from the central split).
}
\examples{
# Load returns of assets or portfolios
data("Industry_10")
rets <- Industry_10
sigma <- cov(rets)
HRP(sigma, UB = 0.15, tau = 0.5)

}
\references{
Lopez de Prado, M. (2016).
Building Diversified Portfolios that Outperform Out-of-Sample.
\emph{SSRN Electronic Journal}.

Pfitzinger, J., Katzke, N. (2019).
A Constrained Hierarchical Risk Parity Algorithm with Cluster-Based Capital Allocation.
\emph{Stellenbosch University, Department of Economics}. Working Paper 14/2019.
}
\author{
Johann Pfitzinger
}
