% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MV.R
\name{MV}
\alias{MV}
\title{Mean Variance}
\usage{
MV(
  sigma,
  mu = NULL,
  UB = NULL,
  LB = NULL,
  groups = NULL,
  group.UB = NULL,
  group.LB = NULL,
  groups_mat = NULL,
  gamma = 0
)
}
\arguments{
\item{sigma}{a \eqn{(N \times N)}{(N x N)} covariance matrix.}

\item{mu}{a \eqn{(N \times 1)}{(N x 1)} vector of estimated returns.}

\item{UB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of upper bound weight constraint.}

\item{LB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of lower bound weight constraint.}

\item{groups}{vector of group IDs. The names of the vector must be identical to the asset names.}

\item{group.UB}{scalar or \eqn{(N_groups\times 1)}{(N_groups x 1)} vector of upper bound group constraints.}

\item{group.LB}{scalar or \eqn{(N_groups\times 1)}{(N_groups x 1)} vector of lower bound group constraints.}

\item{gamma}{risk aversion parameter. Default: \code{gamma = 0}.}
}
\value{
A \eqn{(N \times 1)}{(N x 1)} vector of optimal portfolio weights.
}
\description{
Computes a Mean Variance portfolio with full investment and weight constraints.
}
\details{
The argument \code{sigma} is a covariance matrix.

The MV solution is calculated using \code{quadprog}. When \code{gamma} is left at the
default setting, the minimum variance portfolio is computed.
}
\examples{
# Load returns of assets or portfolios
data("Industry_10")
rets <- Industry_10
sigma <- cov(rets)
MV(sigma, UB = 0.15)

}
\author{
Johann Pfitzinger
}
