% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEW.R
\name{CEW}
\alias{CEW}
\title{Clustered Equal Weights}
\usage{
CEW(sigma, UB = NULL, LB = NULL, n_clusters = "auto", ...)
}
\arguments{
\item{sigma}{a \eqn{(N \times N)}{(N x N)} covariance matrix.}

\item{UB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of upper bound weight constraint.}

\item{LB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of lower bound weight constraint.}

\item{n_clusters}{trade-off between naive (0) or cluster-based (1) tree-splitting (see Details).}

\item{...}{arguments passed to \code{cluster::agnes} method.}
}
\value{
A \eqn{(N \times 1)}{(N x 1)} vector of optimal portfolio weights.
}
\description{
Computes optimal CEW portfolio with full investment and weight constraints.
}
\details{
The argument \code{sigma} is a covariance matrix.

Hierarchical clustering is performed using the \code{cluster}-package. If
\code{cluster_method == 'DIANA'}, the function \code{cluster::diana} is used
to compute a cluster dendrogram, otherwise the function \code{cluster::agnes(., method = cluster_method)}
is used. Default is single-linkage agglomerative nesting.
The number of clusters can be passed using the \code{n_clusters} argument,
calculated automatically with \code{n_clusters='auto'} using the Silhouette criterion.

CEW calculates within cluster minimum variance portfolios coupled with between cluster equal-weights.
Constraints are implemented using an iterative convergence algorithm. Between cluster weights
can deviate from equal weights if the equally weighted solution does not satisfy the box constraints.
}
\examples{
# Load returns of assets or portfolios
data("Industry_10")
rets <- Industry_10
sigma <- cov(rets)
CEW(sigma, UB = 0.15, n_clusters = 'auto')

}
\references{

}
\author{
Johann Pfitzinger
}
