% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSigma.R
\name{chiSigma}
\alias{chiSigma}
\title{A convex hierarchical filter of the covariance matrix.}
\usage{
chiSigma(
  sigma,
  mu = NULL,
  cluster_method = c("single", "average", "complete", "ward", "DIANA"),
  meta_loss = c("MaxDiv", "ERC"),
  gamma = 0
)
}
\arguments{
\item{sigma}{a \eqn{(N \times N)}{(N x N)} covariance matrix.}

\item{mu}{a \eqn{(N \times 1)}{(N x 1)} vector of estimated returns.}

\item{cluster_method}{hierarchical cluster algorithm used to construct an asset hierarchy.}

\item{meta_loss}{a loss function of the most diversified hierarchical allocation graph.}

\item{gamma}{risk aversion parameter. Default: \code{gamma = 0}.}
}
\value{
A \eqn{(N \times N)}{(N x N)} filtered covariance matrix.
}
\description{
Calculates covariance matrix filtered using the most diversified hierarchical graph.
The hierarchical graph can be optimized using maximum diversification (MaxDiv) or equal risk contribution (ERC).
}
\details{
The argument \code{sigma} is a covariance matrix.

Hierarchical clustering is performed using the \code{cluster}-package. If
\code{cluster_method == 'DIANA'}, the function \code{cluster::diana} is used
to compute a cluster dendrogram, otherwise the function \code{cluster::agnes(., method = cluster_method)}
is used. Default is single-linkage agglomerative nesting.

The argument \code{meta_loss} represents the loss function used to optimize the most diversified hierarchical allocation graph.
The optimized hierarchy is used to filter \code{sigma}. If the filtered covariance matrix is used in a
minimum variance portfolio optimizer, a CHI portfolio is constructed.
}
\examples{
# Load returns of assets or portfolios
data("Industry_10")
rets <- Industry_10
sigma <- cov(rets)
chiSigma(sigma)

}
\references{

}
\author{
Johann Pfitzinger
}
