% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hSigma.R
\name{hSigma}
\alias{hSigma}
\title{Hierarchical filtering of the covariance matrix}
\usage{
hSigma(
  sigma,
  cluster_method = c("single", "average", "complete", "ward", "DIANA")
)
}
\arguments{
\item{sigma}{a \eqn{(N \times N)}{(N x N)} covariance matrix.}

\item{cluster_method}{hierarchical cluster algorithm used to construct an asset hierarchy.}
}
\value{
A \eqn{(N \times N)}{(N x N)} filtered covariance matrix.
}
\description{
Generates a hierarchically filtered covariance matrix than can be used for optimization.
}
\details{
The argument \code{sigma} is a covariance matrix.

Hierarchical clustering is performed using the \code{cluster}-package. If
\code{cluster_method == 'DIANA'}, the function \code{cluster::diana} is used
to compute a cluster dendrogram, otherwise the function \code{cluster::agnes(., method = cluster_method)}
is used. Default is single-linkage agglomerative nesting.
}
\examples{
# Load returns of assets or portfolios
data("Industry_10")
rets <- Industry_10
sigma <- cov(rets)
hsig <- hSigma(sigma)
MV(hsig)

}
\references{

}
\author{
Johann Pfitzinger
}
