% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MV.R
\name{MV}
\alias{MV}
\title{Minimum Variance}
\usage{
MV(sigma, UB = NULL, LB = NULL)
}
\arguments{
\item{sigma}{a \eqn{(N \times N)}{(N x N)} covariance matrix.}

\item{UB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of upper bound weight constraint.}

\item{LB}{scalar or \eqn{(N\times 1)}{(N x 1)} vector of lower bound weight constraint.}
}
\value{
A \eqn{(N \times 1)}{(N x 1)} vector of optimal portfolio weights.
}
\description{
Computes a Minimum Variance portfolio with full investment and weight constraints.
}
\details{
The argument \code{sigma} is a covariance matrix.

The MV solution is calculated using \code{quadprog}.
}
\examples{
# Load returns of assets or portfolios
data("Industry_10")
rets <- Industry_10
sigma <- cov(rets)
MV(sigma, UB = 0.15)

}
\references{

}
\author{
Johann Pfitzinger
}
